Refactor task repository structure and implement TaskRepositoryImpl
