+++ src/main/kotlin/di/RepositoryModule.kt
@@ -1,6 +1,6 @@
 package org.baghdad.di
 
-import org.baghdad.data.repository.task.TaskRepositoryImpl
+import org.baghdad.data.repositories.task.TaskRepositoryImpl
 import org.baghdad.logic.repositories.TaskRepository
 import org.koin.dsl.module
